Split import-export into multiple files
